Code that solves the classification problems.
We started by reading in the dataset, preprocessing it, exploring it to find useful insights, and finally building various machine learning models and evaluating them. The main objective of this project was to analyze the titanic dataset and predict whether a passenger will survive or not, based on various input features.

Finally many thanks to Dr.Mohamed Hassan for his efforts alongside the semester.